requirements.txt
fastapi
uvicorn
textblob
python-dotenv
nltk
